requirements.txt
streamlit>=1.30.0
psycopg2-binary>=2.9.9
pandas>=2.0.0
plotly>=5.18.0
sqlparse>=0.4.4
pydantic>=2.5.0
python-dotenv>=1.0.0
pyarrow>=14.0.0
